// src/entities/enums.ts
export enum LoadingStatus {
  Default = "default",
  Loading = "loading",
  Loaded = "loaded",
  Error = "error",
}

export enum ResponseKind {
  Success = "success",
  Failure = "failure",
}

export enum PageNames {
  MainPage = "MainPage",
  NewsPage = "NewsPage",
}

export enum LabelText {
  Refresh = "ОБНОВИТЬ",
  Error = "Что-то пошло не так...",
  AllNews = "Главная",
  FavouriteNews = "Избранное",
  Title = "Новости Альфа-банка",
  EmptyView = "Новостей нет...",
}